requirements.txt
rivertils>=0.1.13
spotipy>=2.21.0
python-decouple>=3.6
python-dateutil>=2.8.2
python-dotenv>=0.21.0
rich>=12.6.0
gspreader>=0.1.20
gspread>=5.7.2
Flask>=3.0.3
Flask-Cors>=4.0.1
Flask-Migrate>=4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn>=22.0.0
requests>=2.26.0

psycopg2==2.9.9